requirements.txt
# Core Dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3

# NLP Libraries (simplified for easier installation)
nltk==3.8.1
transformers==4.35.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.9.0

# LLM Integration
google-generativeai==0.3.2
openai==1.3.7
anthropic==0.7.7

# Utilities
python-dotenv==1.0.0

# UI Components
plotly==5.17.0
matplotlib==3.8.2

# Optional: spaCy (install manually if needed)
# spacy==3.7.2
# Run: pip install spacy && python -m spacy download en_core_web_sm